+++ apps-client/mobile/components/playground.tsx
@@ -1,5 +1,6 @@
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react'
 import { Platform, Pressable, ScrollView, TextProps, View } from 'react-native'
+import { AccessToken, LoginManager } from 'react-native-fbsdk-next'
 import ImageView from 'react-native-image-viewing'
 import { useSafeAreaInsets } from 'react-native-safe-area-context'
 
@@ -40,6 +41,7 @@ import { useEvent } from 'expo'
 import { Image } from 'expo-image'
 import { useRouter } from 'expo-router'
 import { getItemAsync } from 'expo-secure-store'
+import { useTrackingPermissions } from 'expo-tracking-transparency'
 import { useVideoPlayer, VideoView } from 'expo-video'
 import LottieView from 'lottie-react-native'
 import { InfoIcon, PlusIcon, SearchIcon } from 'lucide-react-native'
@@ -60,6 +62,7 @@ export function Playground() {
         <View className="flex flex-row items-center justify-between">
           <H1 className="font-inter-bold">Playground</H1>
         </View>
+        <FacebookSDKExample />
         <View className="flex flex-col gap-2">
           <H2 className="font-inter-bold">Font</H2>
           <View className="flex flex-col gap-1">
@@ -259,6 +262,7 @@ export function Playground() {
         <LottieExample />
         <VideoExample />
         <PostCardExample />
+
         <QueryExample />
         <AuthPlayground />
       </View>
@@ -635,6 +639,100 @@ function PostCardExample() {
         />
         <PostCardSkeleton />
       </View>
+    </View>
+  )
+}
+
+// Settings.initializeSDK()
+
+function FacebookSDKExample() {
+  const [facebookAccessToken, setFacebookAccessToken] = useState<string | null>(null)
+  const facebookPagesQuery = reactQueryClient.useQuery(
+    '/facebook/token/pages',
+    { query: { facebookToken: facebookAccessToken! } },
+    { enabled: !!facebookAccessToken }
+  )
+
+  /**
+   * NOTE: (iOS)
+   * Facebook Login works only with ATTrackingManager enabled
+   * https://github.com/facebook/facebook-ios-sdk/issues/2375#issuecomment-2051743845
+   * https://github.com/thebergamo/react-native-fbsdk-next#troubleshooting (item 9.)
+   */
+  const [status, requestPermission] = useTrackingPermissions()
+  const requestTracking = useCallback(async () => {
+    const { status } = await requestPermission()
+    if (status === 'granted') {
+      console.log('Yay! I have user permission to track data')
+    }
+  }, [requestPermission])
+  return (
+    <View className="flex flex-col gap-2">
+      <H2 className="font-inter-bold">Facebook SDK</H2>
+      {Platform.OS === 'ios' && (
+        <>
+          <Text>Tracking: {JSON.stringify(status, null, 2)}</Text>
+          <Button onPress={requestTracking} variant="outline">
+            <Text>Ask to track</Text>
+          </Button>
+        </>
+      )}
+      {/* <LoginButton
+        onLoginFinished={async (error, result) => {
+          if (error) {
+            console.error('Login failed: ', error)
+            return
+          }
+          console.log('Login success: ' + result)
+          const accessTokenResult = await AccessToken.getCurrentAccessToken()
+          if (!accessTokenResult || !accessTokenResult.accessToken) {
+            console.error('Failed to get facebook access token')
+            return
+          }
+          console.log('AccessToken result: ' + accessTokenResult)
+          setFacebookAccessToken(accessTokenResult.accessToken)
+        }}
+        permissions={['email', 'public_profile']}
+      /> */}
+      <Button
+        disabled={Platform.OS === 'ios' && !status?.granted}
+        onPress={async () => {
+          try {
+            const loginResult = await LoginManager.logInWithPermissions([
+              'pages_show_list',
+              'pages_read_engagement',
+              'pages_read_user_content',
+              'pages_manage_metadata',
+            ])
+            if (loginResult.isCancelled) {
+              console.log('User cancelled login')
+              return
+            }
+            console.log('Login success: ' + loginResult)
+            const accessTokenResult = await AccessToken.getCurrentAccessToken()
+            if (!accessTokenResult || !accessTokenResult.accessToken) {
+              console.error('Failed to get facebook access token')
+              return
+            }
+            console.log('AccessToken result: ' + accessTokenResult)
+            setFacebookAccessToken(accessTokenResult.accessToken)
+          } catch (error) {
+            console.error('Login Error: ', error)
+          }
+        }}
+      >
+        <Text>Login with Facebook</Text>
+      </Button>
+      <Button
+        onPress={() => {
+          LoginManager.logOut()
+        }}
+        variant="secondary"
+      >
+        <Text>Logout</Text>
+      </Button>
+      <Text className="line-clamp-1">Token: {facebookAccessToken}</Text>
+      <Text>Pages: {JSON.stringify(facebookPagesQuery.data, null, 2)}</Text>
     </View>
   )
 }